requirements.txt
pymongo
python-dotenv
pydantic
py-solc-x
web3
cmake
retrying